token services

+++ Server/src/main/java/com/ajparedes/service/ITokenService.java
@@ -4,7 +4,7 @@ import com.ajparedes.model.Token;
 
 public interface ITokenService {
 	
-	public abstract Token getToken(int id);
+	public abstract Token getToken(long id);
 	public abstract Token createToken(String user, String device);
 	public abstract boolean validate(Token token) throws Exception;
